initial cli bootstrapping
