+++ src/Submissions/Api/Commands/Hash.cs
@@ -1,0 +1,13 @@
+using System;
+using System.CommandLine;
+
+namespace Biobanks.Submissions.Api.Commands;
+
+public class Hash : Command
+{
+  public Hash(string commandName)
+    : base(commandName, "A sub-command")
+  {
+    this.SetHandler(() => Console.WriteLine("hello"));
+  }
+}
