+++ src/Submissions/Api/Startup/Cli/CliEntrypoint.cs
@@ -1,0 +1,14 @@
+using System.CommandLine;
+using Biobanks.Submissions.Api.Commands;
+
+namespace Biobanks.Submissions.Api.Startup.Cli;
+
+public class CliEntrypoint : RootCommand
+{
+  public CliEntrypoint() : base("Biobankinguk Directory CLI")
+  {
+    // Add Commands here
+
+    AddCommand(new Hash("hash"));
+  }
+}
